notifications: move email code to common helpers
