+++ backend/kernelCI_app/helpers/email.py
